git-svn-id: http://bwproject.googlecode.com/svn/trunk/bwlib@43 76f41e0e-d1f7-11de-8607-c5ce5d18c61a

+++ bwoss/lin_kmod/bwlib_wrapper.h
@@ -26,7 +26,7 @@
 
 /* bwoss_iface.h */
 typedef spinlock_t bwoss_spinlock_t;
-typedef unsigned long bwoss_irql_t;
+typedef uint32_t bwoss_irql_t;
 static inline void bwoss_spin_lock(bwoss_spinlock_t *l)
 {
 	spin_lock(l);
